+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:datetime_picker_formfield/datetime_picker_formfield.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:intl/intl.dart';
@@ -186,7 +185,8 @@ class MyHomePageState extends State<MyHomePage> {
                     FormBuilderDropdown(
                       attribute: "gender",
                       decoration: InputDecoration(labelText: "Gender"),
-                      // initialValue: 'Male',
+                      initialValue: 'Male',
+                      readonly: true,
                       hint: Text('Select Gender'),
                       validators: [FormBuilderValidators.required()],
                       items: ['Male', 'Female', 'Other']
@@ -283,6 +283,7 @@ class MyHomePageState extends State<MyHomePage> {
                               val >= 10) {
                             return "You can only put more than 10 if you've accepted terms";
                           }
+                          return null;
                         }
                       ],
                     ),
@@ -331,34 +332,34 @@ class MyHomePageState extends State<MyHomePage> {
                               children: languages
                                   .map(
                                     (lang) => Row(
-                                          children: <Widget>[
-                                            Radio<dynamic>(
-                                              value: lang,
-                                              groupValue: field.value,
-                                              onChanged: (dynamic value) {
-                                                field.didChange(lang);
-                                              },
-                                            ),
-                                            lang != "Other"
-                                                ? Text(lang)
-                                                : Expanded(
-                                                    child: Row(
-                                                      children: <Widget>[
-                                                        Text(
-                                                          lang,
-                                                        ),
-                                                        SizedBox(width: 20),
-                                                        Expanded(
-                                                          child: TextFormField(
-                                                            key:
-                                                                _specifyTextFieldKey,
-                                                          ),
-                                                        ),
-                                                      ],
-                                                    ),
-                                                  ),
-                                          ],
+                                      children: <Widget>[
+                                        Radio<dynamic>(
+                                          value: lang,
+                                          groupValue: field.value,
+                                          onChanged: (dynamic value) {
+                                            field.didChange(lang);
+                                          },
                                         ),
+                                        lang != "Other"
+                                            ? Text(lang)
+                                            : Expanded(
+                                                child: Row(
+                                                  children: <Widget>[
+                                                    Text(
+                                                      lang,
+                                                    ),
+                                                    SizedBox(width: 20),
+                                                    Expanded(
+                                                      child: TextFormField(
+                                                        key:
+                                                            _specifyTextFieldKey,
+                                                      ),
+                                                    ),
+                                                  ],
+                                                ),
+                                              ),
+                                      ],
+                                    ),
                                   )
                                   .toList(growable: false),
                             ),