+++ lib/src/fields/form_builder_chips_input.dart
@@ -74,12 +74,12 @@ class _FormBuilderChipsInputState extends State<FormBuilderChipsInput> {
             if (widget.validators[i](val) != null)
               return widget.validators[i](val);
           }
+          return null;
         },
         onSaved: (val) {
           if (widget.valueTransformer != null) {
             var transformed = widget.valueTransformer(val);
-            FormBuilder.of(context)
-                ?.setAttributeValue(widget.attribute, transformed);
+            _formState?.setAttributeValue(widget.attribute, transformed);
           } else
             _formState?.setAttributeValue(widget.attribute, val);
         },