+++ lib/src/fields/form_builder_rate.dart
@@ -64,12 +64,12 @@ class _FormBuilderRateState extends State<FormBuilderRate> {
           if (widget.validators[i](val) != null)
             return widget.validators[i](val);
         }
+        return null;
       },
       onSaved: (val) {
         if (widget.valueTransformer != null) {
           var transformed = widget.valueTransformer(val);
-          FormBuilder.of(context)
-              ?.setAttributeValue(widget.attribute, transformed);
+          _formState?.setAttributeValue(widget.attribute, transformed);
         } else
           _formState?.setAttributeValue(widget.attribute, val);
       },
@@ -91,6 +91,7 @@ class _FormBuilderRateState extends State<FormBuilderRate> {
                     FocusScope.of(context).requestFocus(FocusNode());
                     field.didChange(value);
                     if (widget.onChanged != null) widget.onChanged(value);
+                    return value;
                   },
           ),
         );