+++ lib/src/form_builder.dart
@@ -54,11 +54,9 @@ class FormBuilderState extends State<FormBuilder> {
   }
 
   void setAttributeValue(String attribute, dynamic value) {
-    if (_fieldKeys[attribute] != null) {
-      print("Trying to change value for $attribute to $value");
-      _fieldKeys[attribute].currentState.didChange(value);
-      print(_fieldKeys[attribute].currentState.value);
-    }
+    setState(() {
+      _value[attribute] = value;
+    });
   }
 
   registerFieldKey(String attribute, GlobalKey key) {